Section out and remove components
