--- src/components/Footer/index.tsx
@@ -1,2 +1,0 @@
-export { Footer } from './Footer'
-// export type { FooterProps } from './Footer'
