--- src/components/SideNav/SideNav.types.ts
@@ -1,12 +1,0 @@
-export interface SideNavProps {
-  handleClose?(): void
-}
-
-export interface StyledProps {
-  isOpen: boolean
-}
-
-export interface Props extends SideNavProps, StyledProps {
-  'data-testid'?: string
-  children?: React.ReactNode
-}
