--- src/components/SideNav/SideNavItem/index.ts
@@ -1,2 +1,0 @@
-export { SideNavItem } from './SideNavItem'
-export type { SideNavItemProps } from './SideNavItem'
