--- src/components/SideNav/index.ts
@@ -1,5 +1,0 @@
-export { SideNav } from './SideNav'
-export { SideNavItem } from './SideNavItem'
-
-export type { SideNavProps } from './SideNav'
-export type { SideNavItemProps } from './SideNavItem'
