+++ src/components/index.ts
@@ -5,12 +5,9 @@ export { Hidden }
 export * from './Avatar'
 export * from './Button'
 export * from './Container'
-export * from './Footer'
 export * from './Grid'
-export * from './Header'
 export * from './Heading'
 export * from './Icon'
 export * from './Sidebar'
-export * from './SideNav'
 export * from './Spinner'
 export * from './Toggle'
