+++ src/composites/Footer/index.tsx
@@ -1,0 +1,1 @@
+export { Footer } from './Footer'
