+++ src/composites/page/index.ts
@@ -1,0 +1,2 @@
+export * from './Page'
+export type { PageProps } from './Page'
